""" df starts from 2007-12-04
In the next cell, this will be marked as td = 1, but actually it is td = 3,
since is the 3rd trading day of December 2007.

Overall, it makes little difference in this specific analysis,
but it is better to be accurate if using this data for other purposes since, as an example,
a starting day of 18th of a month would be marked as td = 1, which is not correct.

Due to this, in the next cell, I introduced a filter to keep data a full month
"""